general:   - const correctness fixes core:   - fixed Pow() method in PSPMath (use float version)   - gave parameters in fswrapper open method more comprehensible names audio:   - more robust error handling in audiosystem methods graphics:   - fixed pointer arithmetic issues in PSPHardwareBufferManager   - fixed missing break in D3D11Mappings (detected by ecplipse's codan)

+++ code/core/psp/pspmath.h
@@ -49,7 +49,7 @@ public:
 inline float
 PSPMath::Pow(float base, float exponent)
 {
-    return pow(base, exponent);
+    return powf(base, exponent);
 }
 
 //------------------------------------------------------------------------------
